+++ src/ref_gl/gl_model.h
@@ -247,9 +247,9 @@ uint8_t *Mod_ClusterPVS(int cluster, model_t * model);
 
 void Mod_Modellist_f(void);
 
-void *Hunk_Begin(int maxsize);
-void *Hunk_Alloc(int size);
-int Hunk_End(void);
+void *Hunk_Begin(size_t maxsize);
+void *Hunk_Alloc(size_t size);
+size_t Hunk_End(void);
 void Hunk_Free(void *base);
 
 void Mod_FreeAll(void);
